stop relying on fnamemodify()
This removes the hackery needed to handle all the myriad edge cases of
fnamemodify() and should also make things faster by replacing complex
operations with simpler ones as well as avoiding stat() calls etc.

Ironically this makes the plugin *more* portable rather than less,
because fnamemodify() is such a crappy interface for generic path
manipulation that it’s difficult to handle directory separators well
when using it. Previously I ended up hard-coding slashes in several
places; no more.

The only (potentially) portability-sensitive assumptions the code makes
now is that a path always consists of directories concatenated together
with exactly one single-byte character as a separator, and that for any
path that ends in a directory separator, the :p modifier will append
a trailing separator to it.

The code will work correctly on any system where these assumptions hold.
As far as I know, they are true of every system Vim has been ported to.