use <abuf> to identify buffer during events
This should fix #39 - `lclose` can sometimes cause the tabline to
disappear even when there are multiple buffers remaining. Essentially
the problem is buftabline always assumes the buffer being deleted is the
current buffer, which is generally true, but not always true.

`bufnr('%')` gives us the number of the _current_ buffer which is not
necessarily the buffer that is being added or deleted. `expand(<abuf>)`
gives us the number of buffer on which the event is operating, but as a
string - so we need to convert it to a number with `str2nr`. We can then
determine if the actual buffer being deleted is a user buffer or not.