use <abuf> during BufDelete
Sometimes the tabline disappears even with multiple buffers remaining,
because buftabline assumes that a deletion always targets the current
buffer `bufnr('%')`. This is usually the case, but not always, such as
when invoking `:lclose` from outside the location list.

The buffer being deleted can be queried through `expand('<abuf>')`.

Unfortunately it gets returned as a string, which will never match any
array element unless converted to a number first.

This fixes #39.