ensure identifying newly created quickfix buffers as such
This fixes #75.

The BufAdd event for a quickfix runs before the 'buftype' option on the
buffer has been set, so buftable#user_buffers() cannot detect it for
what it is at that point in time. (All credit to @deathlyfrantic for
tracking this down.)

However, 'filetype' is set shortly after 'buftype' is, and can be relied
on as a signal, since filetype support is not an optional Vim feature.